Zuma Restaurant
The project is mainly based on the client requirements.This application is accessible to everyone want who want to view and purchase our food has always added value of customers everything life.  The day customer able to check the food products details.The client can also check review details about customer feedback. The customers can able to view the products which is  added into the cart of grid view.Contributed to create and write code with backend support .The project was completed within a schedule.